+++ documentation/src/site/sphinx/release_notes.rst
@@ -4,7 +4,8 @@ Release Notes
 
 ${release.version}
 -----
-*
+* Feature     : Maven plugin for running contract server and client
+* Feature     : Fat & Runnable jar for starting ContractServer and running ContractClient
 
 0.0.2
 -----